renamed all three queue files to students instead, example: queue_logic.py to student_logic.py
